+++ com.alibaba.testme.dao.impl.mysql/src/main/java/com/alibaba/testme/dao/impl/SystemDaoImpl.java
@@ -45,7 +45,7 @@ public class SystemDaoImpl extends SqlMapClientDaoSupport implements SystemDao {
      * @return
      */
     @Override
-    public int deleteSystemDO(Integer id) {
+    public int deleteSystemDO(Long id) {
         Integer result = (Integer) this.getSqlMapClientTemplate().delete("system.deleteById", id);
         if (result == null) {
             return 0;
@@ -58,7 +58,7 @@ public class SystemDaoImpl extends SqlMapClientDaoSupport implements SystemDao {
      * @return
      */
     @Override
-    public SystemDO findById(Integer id) {
+    public SystemDO findById(Long id) {
         return (SystemDO) this.getSqlMapClientTemplate().queryForObject("system.findById", id);
     }
