+++ com.alibaba.testme.dao.impl.mysql/src/main/java/com/alibaba/testme/dao/impl/TestunitFlowCaseDaoImpl.java
@@ -47,7 +47,7 @@ public class TestunitFlowCaseDaoImpl extends SqlMapClientDaoSupport implements T
      * @return
      */
     @Override
-    public int deleteTestunitFlowCaseDO(Integer id) {
+    public int deleteTestunitFlowCaseDO(Long id) {
         Integer result = (Integer) this.getSqlMapClientTemplate().delete(
                 "testunitFlowCase.deleteById", id);
         if (result == null) {
@@ -61,7 +61,7 @@ public class TestunitFlowCaseDaoImpl extends SqlMapClientDaoSupport implements T
      * @return
      */
     @Override
-    public TestunitFlowCaseDO findById(Integer id) {
+    public TestunitFlowCaseDO findById(Long id) {
         return (TestunitFlowCaseDO) this.getSqlMapClientTemplate().queryForObject(
                 "testunitFlowCase.findById", id);
     }
