+++ com.alibaba.testme.dao.impl.mysql/src/main/java/com/alibaba/testme/dao/impl/TestunitFlowDetailDaoImpl.java
@@ -48,7 +48,7 @@ public class TestunitFlowDetailDaoImpl extends SqlMapClientDaoSupport implements
      * @return
      */
     @Override
-    public int deleteTestunitFlowDetailDO(Integer id) {
+    public int deleteTestunitFlowDetailDO(Long id) {
         Integer result = (Integer) this.getSqlMapClientTemplate().delete(
                 "testunitFlowDetail.deleteById", id);
         if (result == null) {
@@ -62,7 +62,7 @@ public class TestunitFlowDetailDaoImpl extends SqlMapClientDaoSupport implements
      * @return
      */
     @Override
-    public TestunitFlowDetailDO findById(Integer id) {
+    public TestunitFlowDetailDO findById(Long id) {
         return (TestunitFlowDetailDO) this.getSqlMapClientTemplate().queryForObject(
                 "testunitFlowDetail.findById", id);
     }
