+++ com.alibaba.testme.dao/src/main/java/com/alibaba/testme/dao/TestunitDao.java
@@ -27,13 +27,13 @@ public interface TestunitDao {
      * @param id
      * @return
      */
-    public int deleteTestunitDO(Integer id);
+    public int deleteTestunitDO(Long id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitDO findById(Integer id);
+    public TestunitDO findById(Long id);
 
     /**
      * @param testunitDO
