+++ com.alibaba.testme.service.impl/src/main/java/com/alibaba/testme/service/impl/SystemEnvDetailServiceImpl.java
@@ -49,7 +49,7 @@ public class SystemEnvDetailServiceImpl implements SystemEnvDetailService {
      * @return
      */
     @Override
-    public int deleteSystemEnvDetailDO(Integer id) {
+    public int deleteSystemEnvDetailDO(Long id) {
         if (id == null || id == 0L) {
             return 0;
         }
@@ -61,7 +61,7 @@ public class SystemEnvDetailServiceImpl implements SystemEnvDetailService {
      * @return
      */
     @Override
-    public SystemEnvDetailDO findById(Integer id) {
+    public SystemEnvDetailDO findById(Long id) {
         if (id == null || id == 0L) {
             return null;
         }
