+++ com.alibaba.testme.service.impl/src/main/java/com/alibaba/testme/service/impl/SystemRequirePropServiceImpl.java
@@ -49,7 +49,7 @@ public class SystemRequirePropServiceImpl implements SystemRequirePropService {
      * @return
      */
     @Override
-    public int deleteSystemRequirePropDO(Integer id) {
+    public int deleteSystemRequirePropDO(Long id) {
         if (id == null || id == 0L) {
             return 0;
         }
@@ -61,7 +61,7 @@ public class SystemRequirePropServiceImpl implements SystemRequirePropService {
      * @return
      */
     @Override
-    public SystemRequirePropDO findById(Integer id) {
+    public SystemRequirePropDO findById(Long id) {
         if (id == null || id == 0L) {
             return null;
         }
