+++ com.alibaba.testme.service.impl/src/main/java/com/alibaba/testme/service/impl/TestunitFlowCaseServiceImpl.java
@@ -49,7 +49,7 @@ public class TestunitFlowCaseServiceImpl implements TestunitFlowCaseService {
      * @return
      */
     @Override
-    public int deleteTestunitFlowCaseDO(Integer id) {
+    public int deleteTestunitFlowCaseDO(Long id) {
         if (id == null || id == 0L) {
             return 0;
         }
@@ -61,7 +61,7 @@ public class TestunitFlowCaseServiceImpl implements TestunitFlowCaseService {
      * @return
      */
     @Override
-    public TestunitFlowCaseDO findById(Integer id) {
+    public TestunitFlowCaseDO findById(Long id) {
         if (id == null || id == 0L) {
             return null;
         }
