+++ com.alibaba.testme.service.impl/src/main/java/com/alibaba/testme/service/impl/TestunitFlowServiceImpl.java
@@ -49,7 +49,7 @@ public class TestunitFlowServiceImpl implements TestunitFlowService {
      * @return
      */
     @Override
-    public int deleteTestunitFlowDO(Integer id) {
+    public int deleteTestunitFlowDO(Long id) {
         if (id == null || id == 0L) {
             return 0;
         }
@@ -61,7 +61,7 @@ public class TestunitFlowServiceImpl implements TestunitFlowService {
      * @return
      */
     @Override
-    public TestunitFlowDO findById(Integer id) {
+    public TestunitFlowDO findById(Long id) {
         if (id == null || id == 0L) {
             return null;
         }
