+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitFlowDetailService.java
@@ -27,13 +27,13 @@ public interface TestunitFlowDetailService {
      * @param id
      * @return
      */
-    public int deleteTestunitFlowDetailDO(Integer id);
+    public int deleteTestunitFlowDetailDO(Long id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitFlowDetailDO findById(Integer id);
+    public TestunitFlowDetailDO findById(Long id);
 
     /**
      * @param id
