+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitParamExtService.java
@@ -27,13 +27,13 @@ public interface TestunitParamExtService {
      * @param id
      * @return
      */
-    public int deleteTestunitParamExtDO(Integer id);
+    public int deleteTestunitParamExtDO(Long id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitParamExtDO findById(Integer id);
+    public TestunitParamExtDO findById(Long id);
 
     /**
      * @param id
