+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitService.java
@@ -27,13 +27,13 @@ public interface TestunitService {
      * @param id
      * @return
      */
-    public int deleteTestunitDO(Integer id);
+    public int deleteTestunitDO(Long id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitDO findById(Integer id);
+    public TestunitDO findById(Long id);
 
     /**
      * @param id
