+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/WorkSpaceService.java
@@ -27,13 +27,13 @@ public interface WorkSpaceService {
      * @param id
      * @return
      */
-    public int deleteWorkSpaceDO(Integer id);
+    public int deleteWorkSpaceDO(Long id);
 
     /**
      * @param id
      * @return
      */
-    public WorkSpaceDO findById(Integer id);
+    public WorkSpaceDO findById(Long id);
 
     /**
      * @param id
